test: imageUpload 테스트 작성
