+++ project/__tests__/unit/imageUpload.test.ts
@@ -1,0 +1,6 @@
+import {getThumbnail} from "@/utils/imageUpload";
+
+test('썸네일 추출하기', () => {
+  expect(getThumbnail('![](/__mocks__/fileMock.js)'))
+    .toBe('/__mocks__/fileMock.js');
+});
