+++ project/utils/imageUpload.ts
@@ -6,6 +6,7 @@ import { File, Files } from "formidable";
 const getThumbnail = (markdown: string) => {
   const thumbnailTag = markdown.match(/!\[[^()[\]]*\]\([^)]+\)/)![0];
   const thumbnail = thumbnailTag.match(/\([^\s]*\)?/)![0];
+
   return thumbnail[thumbnail.length - 1] !== ")"
     ? thumbnail.slice(1)
     : thumbnail.slice(1, -1);
